add update student, apply jobs,job assessment,video call integration

+++ app/university/students/student-table/page.tsx
@@ -5,8 +5,10 @@ import React, { useEffect, useState } from "react";
 import Addbutton from "@/components/FormElements/buttons/Addbutton";
 import { useDispatch } from "react-redux";
 import { deleteStudent, fetchAllStudent } from "@/lib/StudentSlice/StudentSlice";
-import { string } from "yup";
 import { DateFilter } from "@/components/Filters/DateFilter/DateFilter";
+import Link from "next/link";
+import { ToastError, ToastSuccess } from "@/components/ToastMessage/ToastMessage";
+import { StudentCourse } from "@/components/Enum/StudentCourse";
 
 interface Student {
   userId: string;
@@ -24,11 +26,13 @@ interface Student {
   createOn: string;
   updateOn: string;
   isApproved: number;
+  course:number;
 }
 
 const AllStudents = () => {
   const dispatch = useDispatch();
   const [students, setStudents] = useState<Student[]>([]);
+
   const fetchData = async () => {
     try {
       const response = await dispatch(fetchAllStudent());
@@ -39,20 +43,29 @@ const AllStudents = () => {
     }
   };
 
-  const onDeleteStudent = async (studentId:string) =>{
+  const onDeleteStudent = async (studentId: string) => {
     try {
+      const confirmed = window.confirm("Are you sure you want to delete this student ?");
+      if (!confirmed) {
+        return;
+      }
+
       const response = await dispatch(deleteStudent(studentId));
-      console.log(response.payload.data); // This should contain the data from your API response
-      response.payload.data && setStudents(response.payload.data);
+      if (response.payload?.success) {
+        ToastSuccess(response.payload?.message);
+        setStudents(prevStudents => prevStudents.filter(student => student.userId !== studentId));
+      } else if (response.error?.message) {
+        ToastError(response.error.message || "An error occurred.");
+      }
     } catch (error) {
       console.error("Error fetching data:", error);
     }
   }
   useEffect(() => {
     fetchData();
-  }, []); // Run once on component mount
+  }, []);
+  // Run once on component mount
 
-  // console.log(students)
   return (
     <>
       <DefaultLayout>
@@ -93,6 +106,9 @@ const AllStudents = () => {
                     <th className="min-w-[120px] px-4 py-4 font-medium text-black dark:text-white">
                       Marital Status
                     </th>
+                    <th className="min-w-[120px] px-4 py-4 font-medium text-black dark:text-white">
+                      Course
+                    </th>
                     <th className="min-w-[150px] px-4 py-4 font-medium text-black dark:text-white">
                       Joining Date
                     </th>
@@ -115,9 +131,9 @@ const AllStudents = () => {
                 </thead>
                 <tbody>
                   {students.map((student, index) => (
-                    <tr key={index}>
+                    <tr key={index + 1}>
                       <td className="border-b border-[#eee] px-4 py-5 pl-9 dark:border-strokedark">
-                        <p className="text-sm">{index}</p>
+                        <p className="text-sm">{index + 1}</p>
                       </td>
                       <td className="border-b border-[#eee] px-4 py-5 dark:border-strokedark">
                         <p className="text-black dark:text-white">
@@ -158,6 +174,11 @@ const AllStudents = () => {
                       </td>
                       <td className="border-b border-[#eee] px-4 py-5 dark:border-strokedark">
                         <p className="text-black dark:text-white">
+                        {StudentCourse[student.course] ?? "--"}
+                        </p>
+                      </td>
+                      <td className="border-b border-[#eee] px-4 py-5 dark:border-strokedark">
+                        <p className="text-black dark:text-white">
                           {DateFilter(student.joiningDate)}
                         </p>
                       </td>
@@ -178,42 +199,29 @@ const AllStudents = () => {
                       </td>
                       <td className="border-b border-[#eee] px-4 py-5 dark:border-strokedark">
                         <p
-                          className={`inline-flex rounded-full bg-opacity-10 px-3 py-1 text-sm font-medium ${
-                            student.isApproved === 1
-                              ? "bg-success text-success"
-                              : student.isApproved === 0
-                                ? "bg-danger text-danger"
-                                : "bg-warning text-warning"
-                          }`}
+                          className={`inline-flex rounded-full bg-opacity-10 px-3 py-1 text-sm font-medium ${student.isApproved == 1
+                            ? "bg-success text-success"
+                            : student.isApproved == 0
+                              ? "bg-danger text-danger"
+                              : "bg-warning text-warning"
+                            }`}
                         >
-                          {student.isApproved === 1
-                            ? "Approved"
-                            : "Not Approved"}
+                          {student.isApproved == 1
+                            ? "Active"
+                            : "Deactive"}
                         </p>
                       </td>
                       <td className="border-b border-[#eee] px-4 py-5 dark:border-strokedark">
                         <div className="flex items-center space-x-3.5">
-                          <button className="hover:text-primary">
-                            <svg
-                              className="fill-current"
-                              width="18"
-                              height="18"
-                              viewBox="0 0 18 18"
-                              fill="none"
-                              xmlns="http://www.w3.org/2000/svg"
-                            >
-                              <path
-                                d="M8.99981 14.8219C3.43106 14.8219 0.674805 9.50624 0.562305 9.28124C0.47793 9.11249 0.47793 8.88749 0.562305 8.71874C0.674805 8.49374 3.43106 3.20624 8.99981 3.20624C14.5686 3.20624 17.3248 8.49374 17.4373 8.71874C17.5217 8.88749 17.5217 9.11249 17.4373 9.28124C17.3248 9.50624 14.5686 14.8219 8.99981 14.8219ZM1.85605 8.99999C2.4748 10.0406 4.89356 13.5562 8.99981 13.5562C13.1061 13.5562 15.5248 10.0406 16.1436 8.99999C15.5248 7.95936 13.1061 4.44374 8.99981 4.44374C4.89356 4.44374 2.4748 7.95936 1.85605 8.99999Z"
-                                fill=""
-                              />
-                              <path
-                                d="M9 11.3906C7.67812 11.3906 6.60938 10.3219 6.60938 9C6.60938 7.67813 7.67812 6.60938 9 6.60938C10.3219 6.60938 11.3906 7.67813 11.3906 9C11.3906 10.3219 10.3219 11.3906 9 11.3906ZM9 7.875C8.38125 7.875 7.875 8.38125 7.875 9C7.875 9.61875 8.38125 10.125 9 10.125C9.61875 10.125 10.125 9.61875 10.125 9C10.125 8.38125 9.61875 7.875 9 7.875Z"
-                                fill=""
-                              />
-                            </svg>
-                          </button>
-                          <button onClick={()=>onDeleteStudent(student.userId)}
-                           className="hover:text-primary">
+                          <Link href={`/university/students/${student.userId}`}>
+                            <button className="hover:text-primary">
+                              <svg className="w-6 h-6 text-gray-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
+                                <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="m14.304 4.844 2.852 2.852M7 7H4a1 1 0 0 0-1 1v10a1 1 0 0 0 1 1h11a1 1 0 0 0 1-1v-4.5m2.409-9.91a2.017 2.017 0 0 1 0 2.853l-6.844 6.844L8 14l.713-3.565 6.844-6.844a2.015 2.015 0 0 1 2.852 0Z" />
+                              </svg>
+                            </button>
+                          </Link>
+                          <button onClick={() => onDeleteStudent(student.userId)}
+                            className="hover:text-primary">
                             <svg
                               className="fill-current"
                               width="18"
